+++ fightext-sprites.ts
@@ -1,0 +1,18 @@
+namespace SpriteKind {
+    export const p1atk = SpriteKind.create()
+    export const p2atk = SpriteKind.create()
+    export const p1body = SpriteKind.create()
+    export const p2body = SpriteKind.create()
+}
+
+namespace fightext_sprites {
+    export enum overlapKind{
+        //% block="敌方弹射物"
+        one,
+        //% block="敌方精灵"
+        two,
+        //% block="任意敌方物体"
+        three
+    }
+
+}
