+++ pxt.json
@@ -13,8 +13,6 @@
         "assets.json",
         "mygame.ts",
         "playgame.ts",
-        "mygame-2.ts",
-        "mygame-1.ts",
         "assets.ts"
     ],
     "testFiles": [
